# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ADYPU Grounded Answer Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/crawl/start:
    post:
      operationId: startCrawl
      tags: [admin]
      summary: Start a new crawl job
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrawlStartRequest"
      responses:
        "200":
          description: Crawl started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrawlJobResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/crawl/recrawl:
    post:
      operationId: recrawlStale
      tags: [admin]
      summary: Recrawl stale content
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recrawl started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrawlJobResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/documents:
    get:
      operationId: getAdminDocuments
      tags: [admin]
      summary: Get indexed documents
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: source
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Documents list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentsListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/questions/low-confidence:
    get:
      operationId: getLowConfidenceQuestions
      tags: [admin]
      summary: Get unanswered or low-confidence questions
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Low-confidence questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LowConfidenceQuestionsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/sources:
    get:
      operationId: getAdminSources
      tags: [admin]
      summary: Get allowed crawl sources
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sources list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourcesListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/sources/add:
    post:
      operationId: addSource
      tags: [admin]
      summary: Add an allowed crawl source
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddSourceRequest"
      responses:
        "200":
          description: Source added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/sources/{id}:
    delete:
      operationId: removeSource
      tags: [admin]
      summary: Remove an allowed crawl source
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Source removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    CrawlStartRequest:
      type: object
      properties:
        sourceIds:
          type: array
          items:
            type: integer
        maxDepth:
          type: integer
          default: 3
    CrawlJobResponse:
      type: object
      properties:
        jobId:
          type: integer
        status:
          type: string
        message:
          type: string
      required:
        - jobId
        - status
    AdminStats:
      type: object
      properties:
        totalDocuments:
          type: integer
        totalChunks:
          type: integer
        failedPages:
          type: integer
        totalSources:
          type: integer
        activeJob:
          $ref: "#/components/schemas/CrawlJobSummary"
        recentJobs:
          type: array
          items:
            $ref: "#/components/schemas/CrawlJobSummary"
        lowConfidenceCount:
          type: integer
      required:
        - totalDocuments
        - totalChunks
        - failedPages
        - totalSources
        - recentJobs
        - lowConfidenceCount
    CrawlJobSummary:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        pagesFound:
          type: integer
        pagesProcessed:
          type: integer
        pagesFailed:
          type: integer
        startedAt:
          type: string
        completedAt:
          type: string
      required:
        - id
        - status
        - pagesFound
        - pagesProcessed
        - pagesFailed
    DocumentsListResponse:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/DocumentItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - documents
        - total
        - page
        - limit
    DocumentItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        sourceUrl:
          type: string
        contentType:
          type: string
        chunkCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - sourceUrl
        - contentType
        - chunkCount
    LowConfidenceQuestionsResponse:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QueryLogItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - questions
        - total
        - page
        - limit
    QueryLogItem:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        intent:
          type: string
        confidenceScore:
          type: number
        confidenceLabel:
          type: string
        createdAt:
          type: string
        hasAnswer:
          type: boolean
      required:
        - id
        - question
        - intent
        - confidenceScore
        - confidenceLabel
        - createdAt
        - hasAnswer
    SourcesListResponse:
      type: object
      properties:
        sources:
          type: array
          items:
            $ref: "#/components/schemas/SourceItem"
      required:
        - sources
    SourceItem:
      type: object
      properties:
        id:
          type: integer
        domain:
          type: string
        urlPattern:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - domain
        - urlPattern
        - isActive
    AddSourceRequest:
      type: object
      properties:
        domain:
          type: string
        urlPattern:
          type: string
        seedUrl:
          type: string
      required:
        - domain
        - urlPattern
...